// src/modules/inventory/reports/yarn-store/yissue-status-report/yarn-issue-status-report-type.ts
export type YarnIssueStatusReportType = {
    REMARKS: string;
    JOB_NO?: string;
    BUYER?: string;
    STYLENO?: string;
    ORDER_NO?: string;
    YARN_CHALLAN_DATE?: string;
    CHALLAN_NO?: string;
    KNITTING_PROGRAM_NO?: string;
    YARN?: string;
    YARN_BRAND?: string;
    YARN_LOT_NUMBER?: string;
    ISSUE_QTY?: number;
    ISSUE_TYPE?: string;
    PARTY?: string;
    COMPANY_NAME?: string;
    COMPANY_ADDRESS?: string;
    SORTING_NO?: string;
    DATA_SOURCE?: string;
    PARTY_TYPE?: string;
}